+++ src/scrapy_redis/connection.py
@@ -1,3 +1,5 @@
+import sys
+
 import six
 
 from scrapy.utils.misc import load_object
@@ -10,8 +12,12 @@ SETTINGS_PARAMS_MAP = {
     'REDIS_URL': 'url',
     'REDIS_HOST': 'host',
     'REDIS_PORT': 'port',
+    'REDIS_DB': 'db',
     'REDIS_ENCODING': 'encoding',
 }
+
+if sys.version_info > (3,):
+    SETTINGS_PARAMS_MAP['REDIS_DECODE_RESPONSES'] = 'decode_responses'
 
 
 def get_redis_from_settings(settings):
@@ -39,10 +45,17 @@ def get_redis_from_settings(settings):
         Server host.
     REDIS_PORT : str, optional
         Server port.
+    REDIS_DB : int, optional
+        Server database
     REDIS_ENCODING : str, optional
         Data encoding.
     REDIS_PARAMS : dict, optional
         Additional client parameters.
+
+    Python 3 Only
+    ----------------
+    REDIS_DECODE_RESPONSES : bool, optional
+        Sets the `decode_responses` kwarg in Redis cls ctor
 
     """
     params = defaults.REDIS_PARAMS.copy()